+++ services/cpts/v1/model/model_history_run_info.go
@@ -12,16 +12,16 @@ type HistoryRunInfo struct {
 	Name *string `json:"name,omitempty"`
 	// run_id
 
-	RunId *int32 `json:"run_id,omitempty"`
+	RunId *float64 `json:"run_id,omitempty"`
 	// run_type
 
-	RunType *int32 `json:"run_type,omitempty"`
+	RunType *float64 `json:"run_type,omitempty"`
 	// start_time
 
 	StartTime *string `json:"start_time,omitempty"`
 	// continue_time
 
-	ContinueTime *int32 `json:"continue_time,omitempty"`
+	ContinueTime *float64 `json:"continue_time,omitempty"`
 	// temp_names
 
 	TempNames *[]TempName `json:"temp_names,omitempty"`
